the code for some tree stuff
I am fairly sure this is np-complete but it is a good exercise. not sure if this is actually how people make graphs but it works and there needs to be code that turns the list of jobs into a directed tree but that doesn't seem so interesting or fun.